pyproject.toml
[project]
name = "mlmodeltraining"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "matplotlib",
    "pillow",
    "torch",
    "torchvision",
    "tqdm",
    "ultralytics",
]
